Fix searching
- `fileStats.isDirectory` is a function, not a property
- Extract function to find test reports into a separate `search.ts` file backed up by tests
