+++ dd-test-results/lib/main.js
@@ -18,44 +18,16 @@ var __importStar = (this && this.__importStar) || function (mod) {
     __setModuleDefault(result, mod);
     return result;
 };
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
 Object.defineProperty(exports, "__esModule", { value: true });
 const core = __importStar(require("@actions/core"));
 const client_1 = require("./client");
 const run_1 = require("./run");
-// eslint-disable-next-line @typescript-eslint/no-require-imports
-const glob = require("@actions/glob");
-const fs_1 = require("fs");
-const util_1 = require("util");
-const stats = util_1.promisify(fs_1.stat);
-function findTestReports(testReportFile) {
-    return __awaiter(this, void 0, void 0, function* () {
-        const globber = yield glob.create(testReportFile);
-        const searchResults = yield globber.glob();
-        const testReportFiles = new Array();
-        for (const searchResult of searchResults) {
-            const fileStats = yield stats(searchResult);
-            if (!fileStats.isDirectory) {
-                core.info(`Test report file found: ${searchResult}`);
-                testReportFiles.push(searchResult);
-            }
-        }
-        return testReportFiles;
-    });
-}
+const search_1 = require("./search");
 run_1.run(new client_1.DataDogClient(core.getInput('dd-api-key', { required: true })), {
     metricName: core.getInput('dd-metric-name', { required: true }),
     tags: core.getMultilineInput('dd-tags', { required: false }),
     host: core.getInput('dd-host', { required: true }),
     testFramework: core.getInput('test-framework', { required: true }),
-    testReportFiles: findTestReports(core.getInput('test-report-file', { required: true })),
+    testReportFiles: search_1.findTestReports(core.getInput('test-report-file', { required: true })),
     testTagsFile: core.getInput('test-tags-file', { required: true })
 });
