+++ dd-test-results/src/main.ts
@@ -1,28 +1,7 @@
 import * as core from '@actions/core'
 import {DataDogClient} from './client'
 import {run} from './run'
-// eslint-disable-next-line @typescript-eslint/no-require-imports
-import glob = require('@actions/glob')
-import {stat} from 'fs'
-import {promisify} from 'util'
-const stats = promisify(stat)
-
-async function findTestReports(testReportFile: string): Promise<string[]> {
-  const globber = await glob.create(testReportFile)
-  const searchResults = await globber.glob()
-  const testReportFiles = new Array<string>()
-
-  for (const searchResult of searchResults) {
-    const fileStats = await stats(searchResult)
-
-    if (!fileStats.isDirectory) {
-      core.info(`Test report file found: ${searchResult}`)
-      testReportFiles.push(searchResult)
-    }
-  }
-
-  return testReportFiles
-}
+import {findTestReports} from './search'
 
 run(new DataDogClient(core.getInput('dd-api-key', {required: true})), {
   metricName: core.getInput('dd-metric-name', {required: true}),
