+++ dd-test-results/tests/search.test.ts
@@ -1,0 +1,8 @@
+import {findTestReports} from '../src/search'
+
+describe('search', () => {
+  it('should find README', async () => {
+    const searchResults = await findTestReports('README.md')
+    expect(searchResults).toContain(process.cwd() + '/README.md')
+  })
+})
